feat :Post에  리액션 추가 로직 옮기기 및 포스트헤더에 갭 주기
