+++ src/components/Header/PostHeader.style.jsx
@@ -91,8 +91,7 @@ export const HeaderServiceContainer = styled.div`
 export const HeaderService = styled.div`
   display: flex;
   padding: 0 2rem;
-  gap: 0.2rem;
-  align-items: flex-start;
+  gap: 2rem;
 
   @media (min-width: 1248px) {
     padding: 0;
